+++ src/main/kotlin/com/ani/todo/discordBot/domain/alarm/repository/AlarmRepository.kt
@@ -9,5 +9,5 @@ interface AlarmRepository : CrudRepository<Alarm, Long> {
     fun findByTitleAndChannelId(title: String, channelId: String): Alarm?
     fun existsByTitleAndChannelId(title: String, channelId: String): Boolean
     fun findByChannelId(channelId: String): List<Alarm>
-    fun findByTime(time: String): List<Alarm>
+    fun findAllByTime(time: String): List<Alarm>
 }
